BigDecimal user instead of Double for price field

+++ src/main/java/mate/academy/internetshop/model/Product.java
@@ -1,12 +1,14 @@
 package mate.academy.internetshop.model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private int count;
 
-    public Product(String productName, Double productPrice, int initialCount) {
+    public Product(String productName, BigDecimal productPrice, int initialCount) {
         name = productName;
         price = productPrice;
         count = initialCount;
@@ -38,11 +40,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
